+++ eslint.config.js
@@ -25,4 +25,11 @@ export default tseslint.config(
       ],
     },
   },
+  // Special configuration for context files
+  {
+    files: ['src/contexts/**/*.{ts,tsx}'],
+    rules: {
+      'react-refresh/only-export-components': 'off',
+    },
+  },
 )
